# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIEngine API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  location?: string | null;
  category: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface Source {
  id: number;
  eventId: number;
  name: string;
  url: string;
  type: string;
  /** @nullable */
  publishedAt?: string | null;
  credibilityScore: number;
  createdAt: string;
}

export interface Claim {
  id: number;
  eventId: number;
  text: string;
  status: string;
  confidence: number;
  createdAt: string;
  updatedAt: string;
}

export interface EventDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  location?: string | null;
  category: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  sources: Source[];
  claims: Claim[];
  trustScore: number;
  trustRating: string;
}

export interface EventList {
  events: Event[];
  total: number;
}

export interface EventInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  summary?: string;
  location?: string;
  category?: string;
}

export interface EventUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  summary?: string;
  location?: string;
  category?: string;
  status?: string;
}

export type SourceList = Source[];

export interface SourceInput {
  eventId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
  type: string;
  publishedAt?: string;
  credibilityScore?: number;
}

export type ClaimList = Claim[];

export interface ClaimInput {
  eventId: number;
  /** @minLength 1 */
  text: string;
  status?: string;
  confidence?: number;
}

export interface Evidence {
  id: number;
  claimId: number;
  /** @nullable */
  sourceId?: number | null;
  text: string;
  strength: string;
  /** @nullable */
  url?: string | null;
  createdAt: string;
}

export type EvidenceList = Evidence[];

export interface EvidenceInput {
  claimId: number;
  sourceId?: number;
  /** @minLength 1 */
  text: string;
  strength: string;
  url?: string;
}

export interface FactCheck {
  id: number;
  claimId: number;
  verdict: string;
  trustScore: number;
  trustRating: string;
  reasoning: string;
  provider: string;
  /** @nullable */
  evidenceSummary?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface FactCheckInput {
  claimId: number;
}

export interface TimelineEntry {
  id: number;
  eventId: number;
  date: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  sourceId?: number | null;
  createdAt: string;
}

export type TimelineList = TimelineEntry[];

export interface SearchResult {
  query: string;
  events: Event[];
  total: number;
}

export type ListEventsParams = {
search?: string;
category?: string;
limit?: number;
offset?: number;
};

export type ListTrendingEventsParams = {
limit?: number;
};

export type ListSourcesParams = {
eventId?: number;
};

export type ListClaimsParams = {
eventId?: number;
status?: string;
};

export type ListEvidenceParams = {
claimId?: number;
};

export type SearchEventsParams = {
q: string;
limit?: number;
};