# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIEngine API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: events
    description: Event operations
  - name: sources
    description: Source operations
  - name: claims
    description: Claim operations
  - name: evidence
    description: Evidence operations
  - name: fact-checks
    description: Fact check operations
  - name: timeline
    description: Timeline operations
  - name: search
    description: Search operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/trending:
    get:
      operationId: listTrendingEvents
      tags: [events]
      summary: List trending events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Trending events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/timeline:
    get:
      operationId: getEventTimeline
      tags: [timeline]
      summary: Get timeline for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineList"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sources:
    get:
      operationId: listSources
      tags: [sources]
      summary: List sources
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceList"
    post:
      operationId: createSource
      tags: [sources]
      summary: Create source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"

  /sources/{id}:
    get:
      operationId: getSource
      tags: [sources]
      summary: Get source by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List claims
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimList"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Create claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"

  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get claim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence
      parameters:
        - name: claimId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceList"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Create evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"

  /fact-check:
    post:
      operationId: runFactCheck
      tags: [fact-checks]
      summary: Run AI fact check on a claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactCheckInput"
      responses:
        "201":
          description: Fact check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactCheck"

  /fact-check/{id}:
    get:
      operationId: getFactCheck
      tags: [fact-checks]
      summary: Get fact check by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fact check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactCheck"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /search:
    get:
      operationId: searchEvents
      tags: [search]
      summary: Search events
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Event:
      type: object
      required: [id, title, description, category, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        summary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    EventDetail:
      type: object
      required: [id, title, description, category, status, createdAt, updatedAt, sources, claims, trustScore, trustRating]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        summary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sources:
          type: array
          items:
            $ref: "#/components/schemas/Source"
        claims:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        trustScore:
          type: number
        trustRating:
          type: string

    EventList:
      type: object
      required: [events, total]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer

    EventInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        summary:
          type: string
        location:
          type: string
        category:
          type: string

    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        summary:
          type: string
        location:
          type: string
        category:
          type: string
        status:
          type: string

    Source:
      type: object
      required: [id, eventId, name, url, type, credibilityScore, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        url:
          type: string
        type:
          type: string
        publishedAt:
          type: ["string", "null"]
          format: date-time
        credibilityScore:
          type: integer
        createdAt:
          type: string
          format: date-time

    SourceList:
      type: array
      items:
        $ref: "#/components/schemas/Source"

    SourceInput:
      type: object
      required: [eventId, name, url, type]
      properties:
        eventId:
          type: integer
        name:
          type: string
          minLength: 1
        url:
          type: string
          minLength: 1
        type:
          type: string
        publishedAt:
          type: string
          format: date-time
        credibilityScore:
          type: integer

    Claim:
      type: object
      required: [id, eventId, text, status, confidence, createdAt, updatedAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        text:
          type: string
        status:
          type: string
        confidence:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ClaimList:
      type: array
      items:
        $ref: "#/components/schemas/Claim"

    ClaimInput:
      type: object
      required: [eventId, text]
      properties:
        eventId:
          type: integer
        text:
          type: string
          minLength: 1
        status:
          type: string
        confidence:
          type: number

    Evidence:
      type: object
      required: [id, claimId, text, strength, createdAt]
      properties:
        id:
          type: integer
        claimId:
          type: integer
        sourceId:
          type: ["integer", "null"]
        text:
          type: string
        strength:
          type: string
        url:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    EvidenceList:
      type: array
      items:
        $ref: "#/components/schemas/Evidence"

    EvidenceInput:
      type: object
      required: [claimId, text, strength]
      properties:
        claimId:
          type: integer
        sourceId:
          type: integer
        text:
          type: string
          minLength: 1
        strength:
          type: string
        url:
          type: string

    FactCheck:
      type: object
      required: [id, claimId, verdict, trustScore, trustRating, reasoning, provider, createdAt, updatedAt]
      properties:
        id:
          type: integer
        claimId:
          type: integer
        verdict:
          type: string
        trustScore:
          type: number
        trustRating:
          type: string
        reasoning:
          type: string
        provider:
          type: string
        evidenceSummary:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    FactCheckInput:
      type: object
      required: [claimId]
      properties:
        claimId:
          type: integer

    TimelineEntry:
      type: object
      required: [id, eventId, date, title, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        date:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: ["string", "null"]
        sourceId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time

    TimelineList:
      type: array
      items:
        $ref: "#/components/schemas/TimelineEntry"

    SearchResult:
      type: object
      required: [query, events, total]
      properties:
        query:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer