# lib/api-zod/src/generated/types/sourceList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIEngine API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Source } from './source';

export type SourceList = Source[];